[Move] useReducer와 관련된 코드를 따로 파일로 만들어 관리
