+++ front/src/component/LogIn.jsx
@@ -1,81 +1,8 @@
 import React, { useReducer } from "react";
 import { Link } from "react-router-dom";
 import styled from "styled-components";
-
-const loginReducer = (state, action) => {
-  switch (action.type) {
-    case 'EMAIL': {
-      return {
-        ...state,
-        email: action.typedEmail,
-      }
-
-    }
-    case 'PASSWORD': {
-      return {
-        ...state,
-        password: action.typedPassword,
-      }
-    }
-    default:
-      return state;
-  }
-}
-
-const enteredLoginDataIsValidReducer = (state, action) => {
-  switch (action.type) {
-    case 'EMAIL': {
-      return {
-        ...state,
-        isEmail: action.isEmailAvailable,
-      }
-
-    }
-    case 'PASSWORD': {
-      return {
-        ...state,
-        isPassword: action.isPasswordAvailable,
-      }
-    }
-    default:
-      return state;
-  }
-}
-
-const isMessageReducer = (state, action) => {
-  switch (action.type) {
-    case 'EMAIL': {
-      return {
-        ...state,
-        emailMessage: action.emailMessage,
-      }
-
-    }
-    case 'PASSWORD': {
-      return {
-        ...state,
-        passwordMessage: action.passwordMessage,
-      }
-    }
-    default:
-      return state;
-  }
-}
-
-const defaultLoginData = {
-  email: "",
-  password: undefined,
-}
-
-const defaultEnteredLoginDataIsValid = {
-  isEmail: false,
-  isPassword: false,
-}
-
-const defaultMessageData = {
-  emailMessage: "",
-  passwordMessage: "",
-}
+import { loginReducer, enteredLoginDataIsValidReducer, isMessageReducer } from "../loginDataHandler/actionReducer";
+import { defaultLoginData, defaultEnteredLoginDataIsValid, defaultMessageData } from "../loginDataHandler/storeData";
