+++ front/src/loginDataHandler/storeData.js
@@ -1,0 +1,17 @@
+const defaultLoginData = {
+    email: "",
+    password: undefined,
+  }
+  
+  const defaultEnteredLoginDataIsValid = {
+    isEmail: false,
+    isPassword: false,
+  }
+  
+  const defaultMessageData = {
+    emailMessage: "",
+    passwordMessage: "",
+  }
+
+
+module.exports = { defaultLoginData, defaultEnteredLoginDataIsValid, defaultMessageData};
